restructuring MyPenalties, MyPermissions 😎😎
Changes to be committed:
	modified:   src/components/AssistenceButton.jsx
	modified:   src/layout/Navigation.jsx
	modified:   src/pages/role.students/MyAssistance/MyAssistance.jsx
	deleted:    src/pages/role.students/MyPenalties.jsx
	new file:   src/pages/role.students/MyPenalties/MyPenalties.jsx
	new file:   src/pages/role.students/MyPenalties/displayUserDetails.js
	new file:   src/pages/role.students/MyPenalties/index.js
	new file:   src/pages/role.students/MyPenalties/tableContent.js
	new file:   src/pages/role.students/MyPenalties/tableHeader.js
	renamed:    src/pages/role.students/MyPermissions.jsx -> src/pages/role.students/MyPermissions/MyPermissions.jsx
	new file:   src/pages/role.students/MyPermissions/MyPermissonTable.jsx
	new file:   src/pages/role.students/MyPermissions/index.js
	new file:   src/pages/role.students/MyPermissions/tableContent.js
	new file:   src/pages/role.students/MyPermissions/tableContentWeekend.js
	new file:   src/pages/role.students/MyPermissions/tableExpand.js
	new file:   src/pages/role.students/MyPermissions/tableHeaderNormal.js
	new file:   src/pages/role.students/MyPermissions/tableheaderWeekend.js
	deleted:    src/pages/role.students/MyPermissonTable.jsx
	modified:   src/routes/preceptorRoutes.js
