+++ frontend/src/pages/role.students/MyPenalties/index.js
@@ -1,0 +1,1 @@
+export { MyPenalties } from './MyPenalties'
