+++ frontend/src/pages/role.students/MyPenalties/tableHeader.js
@@ -1,0 +1,34 @@
+export const tableHeader = [
+  {
+    size: '160px',
+    title: 'Code',
+    display: true,
+    displayMd: true,
+    displaySm: true,
+    color: '#FF719B'
+  },
+  {
+    size: '320px',
+    title: 'Reason',
+    display: true,
+    displayMd: true,
+    displaySm: true,
+    color: '#FF719B'
+  },
+  {
+    size: '120px',
+    title: 'End',
+    display: true,
+    displayMd: false,
+    displaySm: false,
+    color: '#FF719B'
+  },
+  {
+    size: '100px',
+    title: '',
+    display: true,
+    displayMd: true,
+    displaySm: true,
+    color: '#FF719B'
+  }
+]
