+++ frontend/src/pages/role.students/MyPermissions/index.js
@@ -1,0 +1,1 @@
+export { MyPermissions } from './MyPermissions'
