+++ frontend/src/pages/role.students/MyPermissions/tableHeaderNormal.js
@@ -1,0 +1,34 @@
+export const tableHeaderNormal = [
+  {
+    size: '100px',
+    title: 'Type',
+    display: true,
+    displayMd: true,
+    displaySm: true,
+    color: '#B0A3CC'
+  },
+  {
+    size: '300px',
+    title: 'Place',
+    display: true,
+    displayMd: true,
+    displaySm: true,
+    color: '#B0A3CC'
+  },
+  {
+    size: '140px',
+    title: 'Date',
+    display: true,
+    displayMd: true,
+    displaySm: false,
+    color: '#B0A3CC'
+  },
+  {
+    size: '220px',
+    title: 'Out / Entry',
+    display: true,
+    displayMd: false,
+    displaySm: false,
+    color: '#B0A3CC'
+  }
+]
