+++ frontend/src/pages/role.students/MyPermissions/tableheaderWeekend.js
@@ -1,0 +1,34 @@
+export const tableheaderWeekend = [
+  {
+    size: '100px',
+    title: 'Type',
+    display: true,
+    displayMd: true,
+    displaySm: true,
+    color: '#B0A3CC'
+  },
+  {
+    size: '260px',
+    title: 'Place',
+    display: true,
+    displayMd: true,
+    displaySm: true,
+    color: '#B0A3CC'
+  },
+  {
+    size: '180px',
+    title: 'preceptor / vicerector',
+    display: true,
+    displayMd: true,
+    displaySm: false,
+    color: '#B0A3CC'
+  },
+  {
+    size: '220px',
+    title: 'Out / Entry',
+    display: true,
+    displayMd: false,
+    displaySm: false,
+    color: '#B0A3CC'
+  }
+]
