+++ frontend/src/routes/preceptorRoutes.js
@@ -4,7 +4,6 @@ import { Penalties } from '../pages/role.preceptor/Penalties'
 import { MyEvents } from '../pages/role.preceptor/MyEvents'
 import { StudentList } from '../pages/role.preceptor/StudentList'
 import { ValidateWeekends } from '../pages/role.preceptor/ValidateWeekends'
-import { SpecialPermission } from '../pages/role.preceptor/SpecialPermission'
 import { GetPermission } from '../pages/role.students/GetPermission'
 import { ActivateStudent } from '../pages/role.preceptor/ActivateStudent'
